adds settings variables; removes unused import

+++ main.py
@@ -3,6 +3,7 @@ from unicodedata import name
 from fastapi import FastAPI
 
 from app.cache.conn import RedisClientManager
+from app.core.config import settings
 from app.router import api_router
 
 app = FastAPI(debug=True, name="Shortie - Link Shrtnr :)")
@@ -12,6 +13,7 @@ app.include_router(api_router)
 
 @app.on_event("startup")
 async def startup_event():
-    with RedisClientManager() as cache:
-        cache.flushdb()
-        cache.set("counter", 100000000)
+    if settings.ENV == "LOCAL":
+        with RedisClientManager() as cache:
+            cache.flushdb()
+            cache.set(settings.COUNTER_CACHE_KEY, settings.INIT_COUNTER_VALUE)
